+++ crates/cli/src/commands/pin.rs
@@ -3,6 +3,7 @@ use clap::Args;
 use iocraft::prelude::element;
 use proto_core::flow::resolve::Resolver;
 use proto_core::{PinLocation, ProtoConfig, ProtoConfigError, Tool, ToolContext, ToolSpec, cfg};
+use proto_pdk_api::{PinVersionInput, PinVersionOutput, PluginFunction};
 use starbase::AppResult;
 use starbase_console::ui::*;
 use std::path::PathBuf;
@@ -19,8 +20,11 @@ pub struct PinArgs {
     #[arg(long, help = "Resolve the version before pinning")]
     pub resolve: bool,
 
-    #[arg(long, default_value_t, help = "Location of .prototools to pin to")]
+    #[arg(long, default_value_t, help = "Directory location to pin to")]
     pub to: PinLocation,
+
+    #[arg(long, help = "Pin to the tool's native file instead of .prototools")]
+    pub tool_native: bool,
 }
 
 pub async fn internal_pin(
@@ -58,24 +62,79 @@ pub async fn pin(session: ProtoSession, args: PinArgs) -> AppResult {
             .await?;
     }
 
-    let config_path = internal_pin(&tool, &spec, args.to).await?;
+    let config_path;
+
+    if args.tool_native {
+        if tool.plugin.has_func(PluginFunction::PinVersion).await {
+            let output: PinVersionOutput = tool
+                .plugin
+                .call_func_with(
+                    PluginFunction::PinVersion,
+                    PinVersionInput {
+                        context: tool.create_plugin_unresolved_context(),
+                        dir: tool.to_virtual_path(tool.proto.get_config_dir(args.to)),
+                        version: spec.to_unresolved_spec(),
+                    },
+                )
+                .await?;
+
+            if let Some(file) = output.file
+                && output.pinned
+            {
+                config_path = tool.from_virtual_path(file);
+            } else {
+                session.console.render_err(element! {
+                    Notice(variant: Variant::Failure) {
+                        StyledText(
+                            content: format!(
+                                "Failed to pin version <version>{spec}</version> for <id>{}</id>.",
+                                args.context,
+                            )
+                        )
+                        #(output.error.map(|error| {
+                            element! {
+                                StyledText(content: error)
+                            }
+                        }))
+                    }
+                })?;
+
+                return Ok(Some(1));
+            }
+        } else {
+            session.console.render_err(element! {
+                Notice(variant: Variant::Caution) {
+                    StyledText(
+                        content: format!(
+                            "{} does not support pinning to a native file. Remove <shell>--tool-native</shell> and try again.",
+                            tool.get_name()
+                        )
+                    )
+                }
+            })?;
+
+            return Ok(Some(1));
+        }
+    } else {
+        config_path = internal_pin(&tool, &spec, args.to).await?;
+    }
 
     session.console.render(element! {
         Notice(variant: Variant::Success) {
             StyledText(
-                content: if spec != args.spec {
+                content: if args.resolve {
                     format!(
                         "Pinned <id>{}</id> version <version>{}</version> (resolved from <versionalt>{}</versionalt>) to config <path>{}</path>",
                         args.context,
-                        spec,
-                        args.spec,
+                        spec.get_resolved_version(),
+                        spec.req,
                         config_path.display()
                     )
                 } else {
                     format!(
                         "Pinned <id>{}</id> version <version>{}</version> to config <path>{}</path>",
                         args.context,
-                        args.spec,
+                        spec.req,
                         config_path.display()
                     )
                 },